moved Carousel.js into components directory
