+++ Root.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Text, H1, H2, View } from 'native-base'
 import { Link } from 'react-router-native'
 
-import Carousel from "./Carousel"
+import Carousel from './components/Carousel'
 
 export default _ => {
     return (
